fix: split home page component for preview route compatibility
